+++ apps/figma-plugin/src/ui.tsx
@@ -2,7 +2,7 @@ import { render } from '@create-figma-plugin/ui'
 import { h } from 'preact'
 import '!./css/output.css'
 import { tokens } from '@token-icons/core/metadata'
-import { svgs } from '@token-icons/core'
+import { svgs, ITokenMetadata } from '@token-icons/core'
 import { useEffect, useState } from 'preact/hooks'
 import {
   IconCard,
@@ -11,7 +11,6 @@ import {
   ActionBar,
 } from './components'
 import { SvgIcon } from './types'
-import { ITokenMetadata } from '@token-icons/utils'
 
 const PER_PAGE = 50